server/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
flask-bcrypt
flask-socketio
python-dotenv
pymysql
requests
reportlab
cryptography